flask-server/card-generator/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.3
decorator==5.1.1
deep-translator==1.11.4
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.10
Pillow==10.1.0
py==1.11.0
python-dotenv==1.0.0
requests==2.28.2
retry==0.9.2
six==1.16.0
soupsieve==2.5
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2